+++ imml/__init__.py
@@ -1,6 +1,6 @@
 # License: BSD-3-Clause
 
-__version__ = "0.5.1"
+__version__ = "0.6.0"
 
 try:
     import torch