+++ imml/preprocessing/__init__.py
@@ -3,6 +3,7 @@
 from .compose import DropMod, ConcatenateMods, SingleMod, AddMissingMods, SortData, concatenate_mods, drop_mod,\
     single_mod, add_missing_mods, sort_data
 from .mm_transformer import MMTransformer
+from .um_transformer import UMTransformer
 from .select_complete_samples import SelectCompleteSamples, select_complete_samples
 from .normalizer_nan import NormalizerNaN
 from .select_incomplete_samples import SelectIncompleteSamples, select_incomplete_samples